requirements.txt
fastapi==0.98.0
Cython>=0.29.28
cmake>=3.22.3
numpy>=1.21.5
onnxruntime
onnx
tqdm
scikit-image
albumentations
pymysql==1.1.0
opencv-python==4.7.0.72
pillow==9.5.0
faker==18.9.0
bcrypt==4.0.1
sqlalchemy==1.4.39
passlib==1.7.4
scikit-learn==1.2.2
requests==2.31.0
pydantic==1.10.9
fastapi-pagination==0.12.5
python-multipart==0.0.6
uvicorn==0.20.0
websockets==10.4
minio==5.0.1
python-magic-bin==0.4.14
bcrypt==4.0.1
matplotlib==3.7.1
python-jose
cryptography